Programmatically load commands instead of manually
