+++ src/commands/misc/coinFlip.config.ts
@@ -1,5 +1,5 @@
 import { Message, MessageEmbed } from 'discord.js';
-import Command from '../common.commands.config';
+import Command from '../../common.commands.config';
 import config from '../../utils/config';
 
 export default class coinFlipCommand extends Command {
