+++ src/commands/misc/rng.config.ts
@@ -1,6 +1,6 @@
 import { Message, MessageEmbed } from 'discord.js';
 import config from '../../utils/config';
-import Command from '../common.commands.config';
+import Command from '../../common.commands.config';
 
 export default class rngCommand extends Command {
   constructor() {
