+++ src/commands/misc/vote.config.ts
@@ -1,6 +1,6 @@
 import { Message, MessageEmbed, MessageReaction, User } from 'discord.js';
 import config from '../../utils/config';
-import Command from '../common.commands.config';
+import Command from '../../common.commands.config';
 
 export default class voteCommand extends Command {
   constructor() {
